+++ Config/module.config.php
@@ -1,10 +1,13 @@
 <?php
 
 return array(
- 'name'  => 'Menu',
- 'caption'  => 'Menu',
- 'route' => 'Menu:Admin:Item@indexAction',
- 'icon'  => 'fa fa-bars fa-5x',
- 'order' => 3,
- 'description' => 'Menu modules allows you to easily handle different menus on your site'
+    'name'  => 'Menu',
+    'caption'  => 'Menu',
+    'route' => 'Menu:Admin:Item@indexAction',
+    'icon'  => 'fa fa-bars fa-5x',
+    'order' => 3,
+    'description' => 'Menu modules allows you to easily handle different menus on your site',
+    'exclude' => array(
+        // Excluded modules (used when building links) can be set here,
+    )
 );