+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * Ability to exclude modules when building link tree
  * Fixed issue with quote escaping
  * Fixed issue with ability to edit main UL class in `\Menu\View\BootstrapDropdown`
  * Changed color of items. They are now black